refactor: Integrate post files
- Posts로 시작하는 파일들과 Post로 시작하는 파일들이 동시에 존재하여 통합
