+++ src/main/java/project2/repository/PostRepository.java
@@ -1,5 +1,8 @@
 package project2.repository;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,5 +10,5 @@ import project2.entity.Posts;
 
 @Repository
 public interface PostRepository extends JpaRepository<Posts, Long> {
-
+	List<Posts> findByUserUid(Long uid);
 }
